Домашние задания к уроку №9 Переделал на ViewModel
