+++ app/src/main/java/ru/dw/gbkotlinweather/view/contacts/ContactFragment.kt
@@ -1,23 +1,18 @@
 package ru.dw.gbkotlinweather.view.contacts
 
-import android.annotation.SuppressLint
 import android.app.AlertDialog
-import android.content.ContentResolver
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import android.provider.ContactsContract
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
-import ru.dw.gbkotlinweather.R
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
 import ru.dw.gbkotlinweather.databinding.FragmentUserContactBinding
 import ru.dw.gbkotlinweather.model.UserContact
-import ru.dw.gbkotlinweather.utils.TAG
 import ru.dw.gbkotlinweather.utils.arrayPermissions
 import ru.dw.gbkotlinweather.view.contacts.recycler.AdapterContactUser
 import ru.dw.gbkotlinweather.view.contacts.recycler.OnItemListenerContactUser
@@ -26,9 +21,12 @@ import ru.dw.gbkotlinweather.view.contacts.recycler.OnItemListenerContactUser
 class ContactFragment : Fragment(), OnItemListenerContactUser {
     private var _banding: FragmentUserContactBinding? = null
     private val binding get() = _banding!!
+    private val viewModel:ContactUserViewModel by lazy {
+        ViewModelProvider(this).get(ContactUserViewModel::class.java)
+    }
     private var permissionCall = false
     private val adapterContract = AdapterContactUser(this)
-    private val listUserContact = ArrayList<UserContact>()
+
 
 
 
@@ -41,7 +39,7 @@ class ContactFragment : Fragment(), OnItemListenerContactUser {
             if (it.value) {
                 when (it.key) {
                     "android.permission.READ_CONTACTS" -> {
-                        loadPhoneContact()
+                        //loadPhoneContact()
                     }
                     "android.permission.CALL_PHONE" -> {
                         permissionCall = true
@@ -67,7 +65,13 @@ class ContactFragment : Fragment(), OnItemListenerContactUser {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         requestMultiplePermissionLauncher.launch(arrayPermissions)
+        val observer = Observer<List<UserContact>> { data -> adapterContract.submitList(data)}
+        viewModel.getLiveContact().observe(viewLifecycleOwner,observer)
+
+
+
         initRecycler()
+
 
     }
 
@@ -75,30 +79,6 @@ class ContactFragment : Fragment(), OnItemListenerContactUser {
         binding.recyclerViewUserContact.adapter = adapterContract
     }
 
-    @SuppressLint("Range")
-    private fun loadPhoneContact() {
-
-        val contentResolver: ContentResolver = requireActivity().contentResolver
-        val cursor = contentResolver.query(
-            ContactsContract.CommonDataKinds.Phone.CONTENT_URI,
-            null,
-            null,
-            null,
-            null
-        )
-
-        cursor?.let { cur ->
-            while (cur.moveToNext()) {
-                val columnIndexName = cur.getColumnIndex(ContactsContract.Contacts.DISPLAY_NAME)
-                val name = cur.getString(columnIndexName)
-                val phone =
-                    cur.getString(cur.getColumnIndex(ContactsContract.CommonDataKinds.Phone.NUMBER))
-                listUserContact.add(UserContact(name,phone))
-            }
-        }
-        cursor?.close()
-        adapterContract.submitList(listUserContact)
-    }
 
     private fun message(text:String) {
         AlertDialog.Builder(requireContext())
